+++ packages/widgetbook_golden_test_core/lib/src/widgetbook_golden_test_generator.dart
@@ -104,7 +104,7 @@ class WidgetbookGoldenTestGenerator {
           properties: properties,
           useCase: useCase,
           action: play,
-          skip: shouldSkip,
+          skip: play.skip,
           goldenTestBuilder: goldenTestBuilder,
         );
       }