+++ packages/widgetbook_golden_test_core/CHANGELOG.md
@@ -3,9 +3,13 @@
 ### Added
 - Added [precacheImagesTimeout] property to [WidgetbookGoldenTestsProperties] to configure the timeout for precaching images.
 - Added [tags] property to [WidgetbookGoldenTestsProperties] to configure the tags for the golden tests. Also added a [tags] property to [WidgetbookGoldenTestBuilder] to configure the tags per use-case.
+- [GoldenPlayAction] now has a [skip] property to configure the skip per golden play action instead of relying on the [skip] property of [WidgetbookGoldenTestBuilder].
 
 ### Fixed
 - Use runZonedGuarded instead of simple try catch when trying to extract [WidgetbookGoldenTestBuilder] metadata. This solves an issue where no test were being executed when there was an error during the metadata extraction.
+
+### Changed
+- **Breaking Change**: Golden play actions no longer inherit the builder's skip flag by default, each action must set its own skip property if desired.
 
 ## 0.4.0
 Initial release.